fix(quality): corriger le lint frontend, le bundle et la colonne posts.favoris
- Lint : ClubArchiveNav n'exporte plus qu'un composant (fonction getClubPeriods
  et type deplaces dans clubPeriods.ts) ; le reset volontaire vers l'etat de
  chargement avant refetch (ClubPosts, DisciplinePage) est annote et la regle
  react-hooks/set-state-in-effect neutralisee localement avec justification.
  `npm run lint` ne remonte plus aucune erreur.
- Bundle : decoupage des grosses dependances (react, leaflet) via manualChunks ;
  le chunk principal passe de ~541 kB a ~156 kB, l'avertissement > 500 kB
  disparait.
- Base : migration idempotente ajoutant `posts.favoris` si absente (colonne
  utilisee par le modele/l'API mais absente des migrations versionnees). Le test
  du tableau de bord utilise desormais Post::factory(), de nouveau fonctionnelle.
