+++ frontend/src/components/club/ClubArchiveNav.tsx
@@ -1,34 +1,5 @@
 import { Link } from "react-router-dom";
-
-export type ClubPeriod = {
-    label: string;
-    value: string | null;
-};
-
-export function getClubPeriods(
-    currentYear = new Date().getFullYear(),
-): ClubPeriod[] {
-    const currentDecade = Math.floor(currentYear / 10) * 10;
-
-    return [
-        {
-            label: `Depuis ${currentDecade}`,
-            value: `depuis_${currentDecade}`,
-        },
-        {
-            label: `${currentDecade - 10} - ${currentDecade - 1}`,
-            value: `${currentDecade - 10}_${currentDecade - 1}`,
-        },
-        {
-            label: `${currentDecade - 20} - ${currentDecade - 11}`,
-            value: `${currentDecade - 20}_${currentDecade - 11}`,
-        },
-        {
-            label: `Avant ${currentDecade - 20}`,
-            value: `avant_${currentDecade - 20}`,
-        },
-    ];
-}
+import type { ClubPeriod } from "./clubPeriods";
 
 type ClubArchiveNavProps = {
     periods: ClubPeriod[];
